requirements.txt
python-dotenv
bluepy

pytest
pytest-watch

gspread
